Add Binder build trigger in CI (#15)
* Add Binder build trigger for merge to master

On the merge of PRs to the master branch trigger a build of the Binder
for the repo by curl-ing the build API endpoint

* Add connection period and duration limits to Binder build curl

The use of connect-timeout is to make sure that the curl doesn't try to repeatedly hit the URL if the connection is down or unresponsive for an unreasonably long time. The use of max-time is so that for longer builds the curl doesn't try to follow them for the duration. It just hits the API endpoint and then exits at most 60 seconds later.

c.f.
https://unix.stackexchange.com/questions/94604/does-curl-have-a-timeout/94612